chore: moved the function baseQueryWithReauth
